+++ indispensable-josef/src/jvmTest/kotlin/at/asitplus/signum/indispensable/josef/JwsSignedTest.kt
@@ -12,6 +12,7 @@ import io.kotest.core.spec.style.FreeSpec
 import io.kotest.datatest.withData
 import io.kotest.matchers.booleans.shouldBeTrue
 import io.kotest.matchers.nulls.shouldNotBeNull
+import kotlinx.serialization.json.JsonElement
 import java.security.interfaces.RSAPublicKey
 
 class JwsSignedTest : FreeSpec({
@@ -21,7 +22,7 @@ class JwsSignedTest : FreeSpec({
             ?: throw Exception("TestVectors not found")
 
         withData(testvec) { input ->
-            val parsed = JwsSigned.deserialize(input).getOrThrow()
+            val parsed = JwsSigned.deserialize<JsonElement>(input).getOrThrow()
 
             val publicKey = parsed.header.publicKey.shouldNotBeNull()